RAG/requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
openai
python-dotenv
chromadb